[TE] Fixed all compilation errors in thirdeye-spi
