+++ thirdeye/thirdeye-pinot/pom.xml
@@ -175,11 +175,6 @@
 
     <!-- TODO: restructure this -->
     <dependency>
-      <groupId>org.modelmapper</groupId>
-      <artifactId>modelmapper</artifactId>
-      <version>0.7.6</version>
-    </dependency>
-    <dependency>
       <groupId>com.jayway.jsonpath</groupId>
       <artifactId>json-path</artifactId>
     </dependency>
