+++ thirdeye/thirdeye-pinot/src/main/java/org/apache/pinot/thirdeye/alert/fetcher/BaseAnomalyFetcher.java
@@ -21,6 +21,7 @@ package org.apache.pinot.thirdeye.alert.fetcher;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.pinot.thirdeye.alert.commons.AnomalyFetcherConfig;
+import org.apache.pinot.thirdeye.dashboard.resources.v2.AnomaliesResource;
 import org.apache.pinot.thirdeye.datalayer.dto.AlertSnapshotDTO;
 import org.apache.pinot.thirdeye.datalayer.bao.MergedAnomalyResultManager;
 import org.apache.pinot.thirdeye.datalayer.dto.MergedAnomalyResultDTO;
@@ -47,6 +48,10 @@ public abstract class BaseAnomalyFetcher implements AnomalyFetcher {
   protected boolean active = true;
 
   public BaseAnomalyFetcher(){
+  }
+
+  public static String getSnapshotKey(MergedAnomalyResultDTO anomaly) {
+    return anomaly.getMetric() + "::" + AnomaliesResource.generateFilterSetForTimeSeriesQuery(anomaly);
   }
 
   @Override
