+++ thirdeye/thirdeye-pinot/src/main/java/org/apache/pinot/thirdeye/anomaly/detection/AnomalyDetectionInputContextBuilder.java
@@ -32,6 +32,7 @@ import org.apache.pinot.thirdeye.dashboard.Utils;
 import org.apache.pinot.thirdeye.datalayer.dto.AnomalyFunctionDTO;
 import org.apache.pinot.thirdeye.datalayer.dto.DatasetConfigDTO;
 import org.apache.pinot.thirdeye.datalayer.dto.MergedAnomalyResultDTO;
+import org.apache.pinot.thirdeye.datalayer.util.ThirdEyeDataUtils;
 import org.apache.pinot.thirdeye.datasource.DAORegistry;
 import org.apache.pinot.thirdeye.datasource.MetricExpression;
 import org.apache.pinot.thirdeye.datasource.ResponseParserUtils;
@@ -415,7 +416,7 @@ public class AnomalyDetectionInputContextBuilder {
     // Get the original filter
     Multimap<String, String> filters;
     if (StringUtils.isNotBlank(filterString)) {
-      filters = ThirdEyeUtils.getFilterSet(filterString);
+      filters = ThirdEyeDataUtils.getFilterSet(filterString);
     } else {
       filters = HashMultimap.create();
     }
