+++ thirdeye/thirdeye-pinot/src/main/java/org/apache/pinot/thirdeye/detection/algorithm/LegacyDimensionWrapper.java
@@ -23,13 +23,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Multimap;
 import org.apache.pinot.thirdeye.datalayer.dto.AnomalyFunctionDTO;
 import org.apache.pinot.thirdeye.datalayer.dto.DetectionConfigDTO;
+import org.apache.pinot.thirdeye.datalayer.util.ThirdEyeDataUtils;
 import org.apache.pinot.thirdeye.detection.ConfigUtils;
 import org.apache.pinot.thirdeye.detection.DataProvider;
 import org.apache.pinot.thirdeye.detection.DetectionPipeline;
 import org.apache.pinot.thirdeye.detection.DetectionPipelineResult;
 import org.apache.pinot.thirdeye.detector.function.BaseAnomalyFunction;
 import org.apache.pinot.thirdeye.rootcause.impl.MetricEntity;
-import org.apache.pinot.thirdeye.util.ThirdEyeUtils;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -104,7 +104,8 @@ public class LegacyDimensionWrapper extends DimensionWrapper {
 
     if (!properties.containsKey(PROP_METRIC_URN)) {
       long metricId = MapUtils.getLongValue(spec, SPEC_METRIC_ID);
-      Multimap<String, String> filters = ThirdEyeUtils.getFilterSet(MapUtils.getString(spec, SPEC_FILTERS));
+      Multimap<String, String> filters = ThirdEyeDataUtils
+          .getFilterSet(MapUtils.getString(spec, SPEC_FILTERS));
       properties.put(PROP_METRIC_URN,MetricEntity.fromMetric(1.0, metricId, filters).getUrn());
     }
