+++ thirdeye/thirdeye-pinot/src/main/java/org/apache/pinot/thirdeye/detection/wrapper/GrouperWrapper.java
@@ -35,6 +35,7 @@ import org.apache.pinot.thirdeye.detection.DataProvider;
 import org.apache.pinot.thirdeye.detection.DetectionPipeline;
 import org.apache.pinot.thirdeye.detection.DetectionPipelineResult;
 import org.apache.pinot.thirdeye.detection.DetectionUtils;
+import org.apache.pinot.thirdeye.detection.GrouperWrapperConstants;
 import org.apache.pinot.thirdeye.detection.PredictionResult;
 import org.apache.pinot.thirdeye.detection.spi.components.Grouper;
 
@@ -50,7 +51,6 @@ import static org.apache.pinot.thirdeye.detection.yaml.translator.DetectionConfi
 public class GrouperWrapper extends DetectionPipeline {
   private static final String PROP_NESTED = "nested";
   private static final String PROP_GROUPER = "grouper";
-  public static final String PROP_DETECTOR_COMPONENT_NAME = "detectorComponentName";
 
   private final List<Map<String, Object>> nestedProperties;
 
@@ -107,7 +107,7 @@ public class GrouperWrapper extends DetectionPipeline {
       if (anomaly.getProperties() == null) {
         anomaly.setProperties(new HashMap<>());
       }
-      anomaly.getProperties().put(PROP_DETECTOR_COMPONENT_NAME, this.grouperName);
+      anomaly.getProperties().put(GrouperWrapperConstants.PROP_DETECTOR_COMPONENT_NAME, this.grouperName);
       anomaly.getProperties().put(PROP_SUB_ENTITY_NAME, this.entityName);
     }
