+++ thirdeye/thirdeye-spi/pom.xml
@@ -92,6 +92,12 @@
       <scope>compile</scope>
     </dependency>
 
+    <dependency>
+      <groupId>org.modelmapper</groupId>
+      <artifactId>modelmapper</artifactId>
+      <version>0.7.6</version>
+    </dependency>
+
   </dependencies>
 
   <build>
