+++ thirdeye/thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/datalayer/dto/AlertSnapshotDTO.java
@@ -23,11 +23,7 @@ package org.apache.pinot.thirdeye.datalayer.dto;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
 import org.apache.pinot.thirdeye.alert.commons.AnomalyNotifiedStatus;
-import org.apache.pinot.thirdeye.common.time.TimeGranularity;
-import org.apache.pinot.thirdeye.dashboard.resources.v2.AnomaliesResource;
-import org.apache.pinot.thirdeye.datalayer.bao.MergedAnomalyResultManager;
 import org.apache.pinot.thirdeye.datalayer.pojo.AlertSnapshotBean;
-import org.apache.pinot.thirdeye.datasource.DAORegistry;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -35,14 +31,12 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.codehaus.jackson.map.ObjectMapper;
-import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 
 public class AlertSnapshotDTO extends AlertSnapshotBean {
   private static final Logger LOG = LoggerFactory.getLogger(AlertSnapshotDTO.class);
-  public static final TimeGranularity EXPIRE_TIME = TimeGranularity.fromString("7_DAYS");
   private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
   public Multimap<String, AnomalyNotifiedStatus> getSnapshot() {
@@ -81,48 +75,6 @@ public class AlertSnapshotDTO extends AlertSnapshotBean {
     }
   }
 
-  public void updateSnapshot(DateTime alertTime, List<MergedAnomalyResultDTO> alertedAnomalies) {
-    MergedAnomalyResultManager mergedAnomalyResultDAO = DAORegistry.getInstance().getMergedAnomalyResultDAO();
-    // Set the lastNotifyTime to current time if there is alerted anomalies
-    if (alertedAnomalies.size() > 0) {
-      lastNotifyTime = alertTime.getMillis();
-    }
-
-    Multimap<String, AnomalyNotifiedStatus> snapshot = getSnapshot();
-    // update snapshots based on anomalies
-    for (MergedAnomalyResultDTO anomaly : alertedAnomalies) {
-      String snapshotKey = getSnapshotKey(anomaly).toString();
-      AnomalyNotifiedStatus lastNotifyStatus = getLatestStatus(snapshot, snapshotKey);
-      AnomalyNotifiedStatus statusToBeUpdated = new AnomalyNotifiedStatus(alertTime.getMillis(), anomaly.getWeight());
-      if (lastNotifyStatus.getLastNotifyTime() > anomaly.getStartTime()) {
-        // anomaly is a continuing issue, and the status should be appended to the end of snapshot
-        snapshot.put(snapshotKey, statusToBeUpdated);
-      } else {
-        // anomaly is a new issue, override the status list
-        snapshot.removeAll(snapshotKey);
-        snapshot.put(snapshotKey, statusToBeUpdated);
-      }
-
-      // Set notified flag
-      anomaly.setNotified(true);
-      mergedAnomalyResultDAO.update(anomaly);
-    }
-
-    // cleanup stale status in snapshot
-    DateTime expiredTime = alertTime.minus(EXPIRE_TIME.toPeriod());
-    List<String> keysToBeRemoved = new ArrayList<>();
-    for (String snapshotKey : snapshot.keySet()) {
-      AnomalyNotifiedStatus latestStatus = getLatestStatus(snapshot, snapshotKey);
-      if (latestStatus.getLastNotifyTime() < expiredTime.getMillis()) {
-        keysToBeRemoved.add(snapshotKey);
-      }
-    }
-    for (String key : keysToBeRemoved) {
-      snapshot.removeAll(key);
-    }
-    setSnapshot(snapshot);
-  }
-
   public AnomalyNotifiedStatus getLatestStatus(Multimap<String, AnomalyNotifiedStatus> snapshot, String statusKey) {
     if (!snapshot.containsKey(statusKey)) {
       return new AnomalyNotifiedStatus(0, 0);
@@ -139,9 +91,5 @@ public class AlertSnapshotDTO extends AlertSnapshotBean {
     }
 
     return recentStatus;
-  }
-
-  public static String getSnapshotKey(MergedAnomalyResultDTO anomaly) {
-    return anomaly.getMetric() + "::" + AnomaliesResource.generateFilterSetForTimeSeriesQuery(anomaly);
   }
 }
