+++ thirdeye/thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/datalayer/util/ThirdEyeDataUtils.java
@@ -1,0 +1,78 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ *
+ */
+
+package org.apache.pinot.thirdeye.datalayer.util;
+
+import com.google.common.collect.ArrayListMultimap;
+import com.google.common.collect.Multimap;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Set;
+import org.apache.commons.lang3.StringUtils;
+
+public class ThirdEyeDataUtils {
+
+  public static final String FILTER_VALUE_ASSIGNMENT_SEPARATOR = "=";
+  public static final String FILTER_CLAUSE_SEPARATOR = ";";
+
+  public static Multimap<String, String> getFilterSet(String filters) {
+    Multimap<String, String> filterSet = ArrayListMultimap.create();
+    if (StringUtils.isNotBlank(filters)) {
+      String[] filterClauses = filters.split(FILTER_CLAUSE_SEPARATOR);
+      for (String filterClause : filterClauses) {
+        String[] values = filterClause.split(FILTER_VALUE_ASSIGNMENT_SEPARATOR, 2);
+        if (values.length != 2) {
+          throw new IllegalArgumentException("Filter values assigments should in pairs: " + filters);
+        }
+        filterSet.put(values[0], values[1]);
+      }
+    }
+    return filterSet;
+  }
+
+  public static String getSortedFiltersFromMultiMap(Multimap<String, String> filterMultiMap) {
+    Set<String> filterKeySet = filterMultiMap.keySet();
+    ArrayList<String> filterKeyList = new ArrayList<String>(filterKeySet);
+    Collections.sort(filterKeyList);
+
+    StringBuilder sb = new StringBuilder();
+    for (String filterKey : filterKeyList) {
+      ArrayList<String> values = new ArrayList<String>(filterMultiMap.get(filterKey));
+      Collections.sort(values);
+      for (String value : values) {
+        sb.append(filterKey);
+        sb.append(FILTER_VALUE_ASSIGNMENT_SEPARATOR);
+        sb.append(value);
+        sb.append(FILTER_CLAUSE_SEPARATOR);
+      }
+    }
+    return StringUtils.chop(sb.toString());
+  }
+
+  public static String getSortedFilters(String filters) {
+    Multimap<String, String> filterMultiMap = getFilterSet(filters);
+    String sortedFilters = getSortedFiltersFromMultiMap(filterMultiMap);
+
+    if (StringUtils.isBlank(sortedFilters)) {
+      return null;
+    }
+    return sortedFilters;
+  }
+}
