+++ thirdeye/thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/detection/spec/AbstractSpec.java
