+++ thirdeye/thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/detection/spi/components/BaseComponent.java
