+++ thirdeye/thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/detection/spi/model/EvaluationSlice.java
