+++ thirdeye/thirdeye-spi/src/main/java/org/apache/pinot/thirdeye/rootcause/impl/TimeRangeEntity.java
